+++ src/query-builder/components/__tests__/__mocks__/configure_annotation_ev.json
@@ -1,0 +1,56 @@
+[
+  {
+    "groupName": "Any",
+    "items": [
+      {
+        "name": "Any assertion method",
+        "code": "any"
+      },
+      {
+        "name": "Any manual assertion",
+        "code": "manual"
+      },
+      {
+        "name": "Any automatic assertion",
+        "code": "automatic"
+      },
+      {
+        "name": "Any experimental assertion",
+        "code": "experimental"
+      }
+    ]
+  },
+  {
+    "groupName": "Manual assertions",
+    "items": [
+      {
+        "name": "Experimental",
+        "code": "ECO_0000269"
+      },
+      {
+        "name": "Non-traceable author statement",
+        "code": "ECO_0000303"
+      },
+      {
+        "name": "Curator inference",
+        "code": "ECO_0000305"
+      },
+      {
+        "name": "Sequence similarity",
+        "code": "ECO_0000250"
+      },
+      {
+        "name": "Sequence model",
+        "code": "ECO_0000255"
+      },
+      {
+        "name": "Combinatorial",
+        "code": "ECO_0000244"
+      },
+      {
+        "name": "Imported information",
+        "code": "ECO_0000312"
+      }
+    ]
+  }
+]
